fix(review): 实施 2026-08-29 深度审查修复（计数门禁/架构回归/安全加固/测试CI）
Phase 1 计数与门禁：
- 232→236 散文计数全部同步（AGENTS/CONTRIBUTING/ElementWiseMapper/cross-validation）
- CHANGELOG "8 个函数"→"4 个函数" + 补 [2.2.0] compare 链接 + Unreleased 段
- verify-docs 新增检查 16（散文式 UDF 计数一致性）+ test_verify_docs 场景 G
- pre-commit check-5 名单动态化（含 DOE）+ 除法检测排除注释；check-6 豁免名单收缩为 6 项（与 AGENTS.md 一致）

Phase 2 架构回归：
- ARR.FILL/ARR.RANGE 业务逻辑下沉 ArrayCore（UDF 变单行分发）+ 13 个 Core 测试
- 新增 InputNormalizer.ToInt32（超 int 范围抛异常），替换 Linalg/Pivot/RangeExport/Doe 4 处 (int) 直转
- STATS.COUNT 改 Excel COUNT 语义（CountNumeric 跳过非数值，不再抛 #VALUE!）
- RANGE.TOHTML/TOJSON/TOMD has_headers 可选化；RANGE.TOCSV/PIVOT.* 新增 [has_headers]（api-reference/user-manual 同步）

Phase 3 安全加固：
- FileSystemCore.DeleteFolderRecursive 改迭代栈式遍历（防未捕获 StackOverflow）
- SQLite 原生 DLL 抽取改 SHA-256 完整性校验（不再仅比大小）
- AutoOpen 沙箱未启用时输出醒目警告（决策点留待产品确认默认受限根）

Phase 4 测试与 CI：
- ErrorMsg/ExceptionFilters/ExcelEmpty 补专用测试（Foundation +18 测试）
- DOE.ANALYZE/ANOVA/PARETO 接入 Python 交叉验证（Dispatcher 注册 + manifest + verify-manual 独立实现，3 项全匹配）
- security.yml 增加 pull_request 触发（P2-9）
- release/csproj 打包验证改硬错误：去除 ContinueOnError + Warning→Error + 空产物断言（P2-10）
- update_excel_arguments.py 路径修复 rules/api-reference.md（P2-8）

验证：dotnet test 6 项目全绿（325+684+1395）×2 TFM；verify-manual 328 项 0 失败；verify-docs 20/20；pre-commit 6/6；Release 构建 0 警告

+++ src/Analytics/LinalgUdf.cs
@@ -78,7 +78,7 @@ namespace ExcelFormulaLabs.Analytics
 
         [ExcelFunction(Name = "LINALG.IDENTITY", Description = "Identity matrix.")]
         public static object UDF_LINALG_IDENTITY([ExcelArgument(Name="size", Description="The matrix dimension, e.g. 3 for a 3x3 identity matrix")] object n)
-            => OutputWrapper.WrapError(() => LinalgCore.Identity((int)InputNormalizer.ToLong(n)));
+            => OutputWrapper.WrapError(() => LinalgCore.Identity(InputNormalizer.ToInt32(n)));
 
         [ExcelFunction(Name = "LINALG.MATMUL", Description = "Matrix multiplication.")]
         public static object UDF_LINALG_MATMUL([ExcelArgument(Name="array1", Description="First range or array")] object A, [ExcelArgument(Name="array2", Description="Second range or array")] object B)
